fix(#1092): keep types.ts under size cap + scope admin-portal nav-leak e2e
CI on PR #1137 surfaced two failures the local run missed:
- lint:size: the inline AdminBookingFilters tipped the types.ts barrel to
  805 lines (cap 800). Extracted it to types-admin-booking.ts and re-export
  from the barrel, mirroring the consent/payment/review split. Barrel now 784.
- e2e admin-portal nav-leak: the new (visible) AdminSidebar Bookings link
  collided with the renter global-nav Bookings sentinel. Scoped the
  toBeHidden() assertion to [data-global-nav] so it tests the global-nav leak
  only, preserving the #462/#541 security intent.
